Foi corrigido o carregamento da pagina músico e juntou- se a classe de Review com as classes das outras 2 abas.

+++ app/src/main/java/com/rogacheski/lbd/lbdmusic/entity/ReviewsEntity.java
@@ -7,7 +7,7 @@ import java.util.Date;
  */
 
 public class ReviewsEntity {
-    int iGrade;
+    float iGrade;
     String description;
     Date dData_Review;
     String sEvaluators_Name;
@@ -22,11 +22,11 @@ public class ReviewsEntity {
     }
 
 
-    public int getiGrade() {
+    public float getiGrade() {
         return iGrade;
     }
 
-    public void setiGrade(int iGrade) {
+    public void setiGrade(float iGrade) {
         this.iGrade = iGrade;
     }
